register 'Source' and include in rest api
WordPress only includes meta fields that are explicitly declared to be
shown in the REST API. This can be done my calling `register_meta` on
the object type, which is 'post' in this case, and setting the option
`show_in_rest` to true.

I've also specified the value to be a single, string type.

Documentation: https://developer.wordpress.org/reference/functions/register_meta/
Fixes: #1